// src/api/stetsom/model/cmsProductRowStatus.ts
/**
 * Generated by orval v8.10.0 🍺
 * Do not edit manually.
 * Stetsom API
 * API documentation for Stetsom
 * OpenAPI spec version: 1.0.0
 */

export type CmsProductRowStatus =
  (typeof CmsProductRowStatus)[keyof typeof CmsProductRowStatus];

export const CmsProductRowStatus = {
  PUBLISHED: "PUBLISHED",
  DRAFT: "DRAFT",
  SCHEDULED: "SCHEDULED",
} as const;